- add support of version 3.1.X of gatling - adjust percentile percentages to NFR - add geometricMean calculations - display milliseconds in startDate - removed apdex from report - fixed request start time to not use Simulation start time and use min request start time - fixed calculation of request duration

+++ src/main/java/org/nuxeo/tools/gatling/report/TrendContext.java
@@ -90,7 +90,7 @@ public class TrendContext {
                 indice = stat.indice;
                 xvalues.add(String.format("'%s'", stat.startDate));
                 yvalues.add(stat.avg);
-                yerrors.add(stat.stddev);
+                yerrors.add((long) stat.stddev);
                 rps.add(stat.rps);
             }
         }
